+++ tests/Form/Type/StatusTypeTest.php
@@ -80,7 +80,7 @@ class StatusTypeTest extends TypeTestCase
         $options = $resolver->resolve([]);
 
         $this->assertArrayHasKey('choices', $options);
-        $this->assertSame($options['choices'], [1 => 'salut']);
+        $this->assertSame($options['choices'], ['salut' => 1]);
     }
 
     public function testGetDefaultOptionsWithValidFlip(): void
@@ -90,7 +90,7 @@ class StatusTypeTest extends TypeTestCase
             2 => 'toi!',
         ];
 
-        $type = new StatusType(Choice::class, 'getList', 'choice_type', true);
+        $type = new StatusType(Choice::class, 'getList', 'choice_type');
 
         $this->assertSame('choice_type', $type->getName());
         $this->assertSame(ChoiceType::class, $type->getParent());
@@ -112,13 +112,13 @@ class StatusTypeTest extends TypeTestCase
             2 => 'error',
         ];
 
-        $type = new StatusType(Choice::class, 'getList', 'choice_type', true);
+        $type = new StatusType(Choice::class, 'getList', 'choice_type');
 
         $this->assertSame('choice_type', $type->getName());
         $this->assertSame(ChoiceType::class, $type->getParent());
 
         FormHelper::configureOptions($type, $resolver = new OptionsResolver());
 
-        $options = $resolver->resolve([]);
+        $resolver->resolve([]);
     }
 }