+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.9.0](https://github.com/sonata-project/SonataCoreBundle/compare/3.8.0...3.9.0) - 2018-01-07
+### Added
+- Added close label to flash messages
+
+### Changed
+- Always flip choices in `BaseStatusType`
+
 ## [3.8.0](https://github.com/sonata-project/SonataCoreBundle/compare/3.7.1...3.8.0) - 2017-12-16
 ### Added
 - Added Russian translations